Back to serde for now
