+++ asset_common/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+serde = { version = "1.0", features = ["derive"] }
 interprocess = { version = "1.2.1", default-features = false }
-ultraviolet = { version = "0.9.2" }
-rkyv = { version = "0.7", features = ["validation"] }
+bincode = "1.3"
+ultraviolet = { version = "0.9.2", features = ["serde"] }
