+++ asset_common/src/asset/scene.rs
@@ -12,11 +12,11 @@ pub use texture::*;
 
 use std::{borrow::Cow, error::Error};
 
-use rkyv::{Archive, Deserialize, Serialize};
+use serde::{Deserialize, Serialize};
 
 use crate::{AssetData, AssetTypeId};
 
-#[derive(Debug, Archive, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct LoadedScene {
     pub models: Vec<LoadedModel>,
 }
@@ -36,13 +36,10 @@ impl AssetData for LoadedScene {
     }
 
     fn to_bytes(&self) -> Result<Cow<[u8]>, impl Error + 'static> {
-        rkyv::to_bytes::<_, 1024>(self).map(|v| Cow::Owned(v))
+        bincode::serialize(self).map(|v| Cow::Owned(v))
     }
 
     fn from_bytes(bytes: &[u8]) -> Result<Self, impl Error + 'static> {
-        rkyv::check_archived_root::<Self>(bytes)
-            .unwrap()
-            .deserialize(&mut rkyv::Infallible)
-            .unwrap()
+        bincode::deserialize(bytes)
     }
 }
