+++ asset_common/src/asset/scene/gltf_asset.rs
@@ -1,10 +1,12 @@
 use std::{collections::HashMap, sync::Arc};
 
+use serde::{Deserialize, Serialize};
+
 pub trait GltfAsset {
     fn id(&self) -> GltfAssetId;
 }
 
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Deserialize, Serialize)]
 pub struct GltfAssetId(u32);
 impl GltfAssetId {
     pub fn new(id: u32) -> Self {
