+++ asset_server/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 asset_common = { path = "../asset_common" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+bincode = "1.3"
 log = "0.4.20"
 env_logger = "0.10"
 ipc-channel = "0.18"
